Factor out Dropdown component
