+++ client/src/styles/App.scss
@@ -5,18 +5,6 @@
   display: flex;
   flex-direction: column;
 
-  .dropdown-overlay {
-    &.hidden {
-      display: none;
-    }
-    position: fixed;
-    width: 100%;
-    height: 100%;
-    background-color: $dark;
-    opacity: 0.6;
-    z-index: 100;
-  }
-
   & > div {
     box-sizing: border-box;
   }
@@ -55,34 +43,6 @@
 
   @include for-phone-only {
     padding: 10px 8px 8px;
-  }
-  .dropdown.dropdown-right {
-    z-index: 110;
-    // Remove default border on focus:
-    button.dropdown-toggle:focus {
-      border: none;
-    }
-    .menu {
-      padding: 0.25rem 0.5rem;
-      font-size: 1.6rem;
-
-      .menu-item a {
-        &:focus {
-          outline-offset: 0px;
-        }
-        &:active,
-        &.active {
-          background: $gray-light;
-        }
-      }
-
-      &.d-block {
-        display: block;
-      }
-      &.d-none {
-        display: none;
-      }
-    }
   }
 
   .label.label-warning {
