+++ client/src/styles/Dropdown.scss
@@ -1,0 +1,46 @@
+@import "_vars.scss";
+
+.DropdownComponent {
+  .dropdown-overlay {
+    &.hidden {
+      display: none;
+    }
+    position: fixed;
+    top: 0px;
+    right: 0px;
+    width: 100%;
+    height: 100%;
+    background-color: $dark;
+    opacity: 0.6;
+    z-index: 100;
+  }
+
+  .dropdown.dropdown-right {
+    z-index: 110;
+    // Remove default border on focus:
+    button.dropdown-toggle:focus {
+      border: none;
+    }
+    .menu {
+      padding: 0.25rem 0.5rem;
+      font-size: 1.6rem;
+
+      .menu-item a {
+        &:focus {
+          outline-offset: 0px;
+        }
+        &:active,
+        &.active {
+          background: $gray-light;
+        }
+      }
+
+      &.d-none {
+        display: none;
+      }
+      &.d-block {
+        display: block;
+      }
+    }
+  }
+}
